+++ src/main/java/org/cascadebot/cascadebot/commandmeta/CommandContext.java
@@ -8,7 +8,7 @@ package org.cascadebot.cascadebot.commandmeta;
 import org.cascadebot.cascadebot.CascadeBot;
 import org.cascadebot.cascadebot.data.Config;
 import org.cascadebot.cascadebot.data.objects.GuildData;
-import org.cascadebot.cascadebot.data.objects.GuildSettings;
+import org.cascadebot.cascadebot.data.objects.GuildSettingsCore;
 import org.cascadebot.cascadebot.messaging.MessagingDirectMessage;
 import org.cascadebot.cascadebot.messaging.MessagingTimed;
 import org.cascadebot.cascadebot.messaging.MessagingTyped;
@@ -28,17 +28,7 @@ import net.dv8tion.jda.core.entities.User;
 import net.dv8tion.jda.core.utils.Checks;
 import org.apache.commons.lang3.ArrayUtils;
 import org.apache.commons.lang3.StringUtils;
-import org.cascadebot.cascadebot.CascadeBot;
-import org.cascadebot.cascadebot.data.Config;
-import org.cascadebot.cascadebot.data.objects.GuildData;
-import org.cascadebot.cascadebot.data.objects.GuildSettings;
-import org.cascadebot.cascadebot.messaging.MessagingDirectMessage;
-import org.cascadebot.cascadebot.messaging.MessagingTimed;
-import org.cascadebot.cascadebot.messaging.MessagingTyped;
-import org.cascadebot.cascadebot.messaging.MessagingUI;
 import org.cascadebot.cascadebot.music.CascadePlayer;
-import org.cascadebot.cascadebot.permissions.CascadePermission;
-import org.cascadebot.shared.Regex;
 
 import java.util.HashSet;
 import java.util.Set;
@@ -83,7 +73,7 @@ public class CommandContext {
         return data;
     }
 
-    public GuildSettings getSettings() {
+    public GuildSettingsCore getSettings() {
         return data.getSettings();
     }
