+++ src/main/java/org/cascadebot/cascadebot/commands/subcommands/tag/TagListSubCommand.java
@@ -18,7 +18,7 @@ public class TagListSubCommand implements ICommandExecutable {
 
     @Override
     public void onCommand(Member sender, CommandContext context) {
-        Map<String, Tag> tags = context.getData().getTagInfo();
+        Map<String, Tag> tags = context.getSettings().getTagInfo();
 
         if (tags.size() == 0) {
             context.getTypedMessaging().replyWarning("Server has no tags!");
